+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Use td-hypermedia 3.6.1
 
+### Deleted
+
+- [TD-2095] Field name from system params' path 
+
 ## [3.5.1] 2019-09-03
 
 ### Fixed